Correcting a defect in SimpleProperty for single-valued properties
