+++ PyOpenWorm/utils.py
@@ -4,7 +4,7 @@ elsewhere in PyOpenWorm
 """
 import re
 
-__all__ = ['normalize_cell_name']
+__all__ = ['normalize_cell_name', 'grouper']
 # to normalize certain neuron and muscle names
 SEARCH_STRING = re.compile(r'\w+0+[1-9]+')
 REPLACE_STRING = re.compile(r'0+')
@@ -29,3 +29,18 @@ def normalize_muscle(name):
     if re.match(SEARCH_STRING_MUSCLE, name):
         name = REPLACE_STRING_MUSCLE.sub('', name)
     return name
+
+def grouper(iterable, n, fillvalue=None):
+    "Collect data into fixed-length chunks or blocks"
+    # grouper('ABCDEFG', 3, 'x') --> ABC DEF Gxx
+    args = [iter(iterable)] * n
+    while True:
+        l = []
+        try:
+            for x in args:
+                l.append(next(x))
+        except Exception:
+            pass
+        yield l
+        if len(l) < n:
+            break
